feat: Use rule of three for addrinfo class
The CAddrinfo class wraps the C like system call getaddrinfo(). This
provides the adddress information with a pointer to an allocated
resource that has to be freed with freeaddrinfo(). It means we have to
manage an external resource with the class and have to observe the rule
of three. Together with the custom destructor to free the external
resource, the class has a copy constructor and and a copy assignment
operator. Instead of copying the resource from the old address
information class we get them again from the operating system. To get
identical values we must use the same hints than before. These hints are
cached and copied not the resource itself because we don't know if there
are additional pointer within the address info that have to be managed
with deep copy.

Also create TRACE2 macro with two arguments.